requirements.txt
pymupdf>=1.24.0
python-docx>=1.2.0
anthropic>=0.30.0
plotly>=5.0.0
jinja2>=3.0.0
python-dotenv>=1.0.0
